+++ upgrade.js
@@ -10,7 +10,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("displayLifespan").classList.remove("hidden"); },
-        purchasetext: "You peer into the unknown. Your lifespan is revealed."
+        purchasetext: "You peer into the unknown. Your lifespan is revealed.",
+        type: "basic",
     },
 
     {
@@ -26,7 +27,8 @@ const upgrades = [
             document.querySelector('[data-tab = "Pit"]').classList.remove("hidden");
             resetThePitTimer();
         },
-        purchasetext: "A big dark hole emerged from the nearby ground. It's deep."
+        purchasetext: "A big dark hole emerged from the nearby ground. It's deep.",
+        type: "basic",
     },
 
     {
@@ -39,7 +41,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("moodContainer").classList.remove("hidden"); },
-        purchasetext: "A bar floats above your head, dictating your mood"
+        purchasetext: "A bar floats above your head, dictating your mood",
+        type: "basic",
     },
 
     {
@@ -54,9 +57,10 @@ const upgrades = [
         onpurchase: () => {
             document.getElementById("displayMatter").classList.remove("hidden");
             document.getElementById("displayMatterBonus").classList.remove("hidden");
-            document.querySelector('[data-tab = "Matter"]').classList.remove("hidden");
+            document.getElementById("matterContainer").classList.remove("hidden");
         },
-        purchasetext: "Matter is shown to ye!"
+        purchasetext: "Matter is shown to ye!",
+        type: "basic",
     },
 
     {
@@ -69,7 +73,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("setToMax").classList.remove("hidden"); },
-        purchasetext: "Click the button to set inputs to player eqivalent (max)"
+        purchasetext: "Click the button to set inputs to player eqivalent (max)",
+        type: "basic",
     },
 
     {
@@ -82,7 +87,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => {},
-        purchasetext: "Instead of on purpose, passively think."
+        purchasetext: "Instead of on purpose, passively think.",
+        type: "basic",
     },
 
     {
@@ -95,7 +101,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => {},
-        purchasetext: "Wisdom requires deep thought. You don't have that but I'm granting you it."
+        purchasetext: "Wisdom requires deep thought. You don't have that but I'm granting you it.",
+        type: "basic",
     },
 
     {
@@ -107,8 +114,12 @@ const upgrades = [
         flavortext: "I may not have a brain gentlemen. But I have an idea..",
         purchased: 0,
         maxpurchases: 1,
-        onpurchase: () => { document.getElementById("vitContainer").classList.remove("hidden"); },
-        purchasetext: "A strict routine only for the best"
+        onpurchase: () => {
+            document.getElementById("vitContainer").classList.remove("hidden");
+            document.querySelector('[data-tab = "Training"]').classList.remove("hidden");
+        },
+        purchasetext: "A strict routine only for the best",
+        type: "basic",
     },
 
     {
@@ -121,7 +132,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("mediateContainer").classList.remove("hidden"); },
-        purchasetext: "Time to get fit!"
+        purchasetext: "Time to get fit!",
+        type: "basic",
     },
 
     {
@@ -134,7 +146,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("martialContainer").classList.remove("hidden"); },
-        purchasetext: "Everybody wa' kung foo fighti-ing hooo haaaa!!"
+        purchasetext: "Everybody wa' kung foo fighti-ing hooo haaaa!!",
+        type: "basic",
     },
 
 
@@ -148,7 +161,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("hitContainer").classList.remove("hidden"); },
-        purchasetext:"You got out of your kung foo phase for the real stuff"
+        purchasetext:"You got out of your kung foo phase for the real stuff",
+        type: "basic",
     },
 
     {
@@ -161,7 +175,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("taichiContainer").classList.remove("hidden"); },
-        purchasetext: "Say a proverb because you don't care about life! PEEEEAAAACE!!!"
+        purchasetext: "Say a proverb because you don't care about life! PEEEEAAAACE!!!",
+        type: "basic",
     },
 
     {
@@ -178,7 +193,8 @@ const upgrades = [
             document.getElementById("displayWisdom").classList.remove("hidden");
             document.getElementById("displayWisdomBonus").classList.remove("hidden");
         },
-        purchasetext: "A white beard sprouts on your chin, and your back teeth emerge. You feel...wiser somehow?"
+        purchasetext: "A white beard sprouts on your chin, and your back teeth emerge. You feel...wiser somehow?",
+        type: "basic",
     },
 
     {
@@ -191,7 +207,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("peaceContainer").classList.remove("hidden"); },
-        purchasetext: "MORE BARS!!!!"
+        purchasetext: "MORE BARS!!!!",
+        type: "basic",
     },
 
     {
@@ -204,7 +221,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("studyContainer").classList.remove("hidden"); },
-        purchasetext: "The deeper the knowledge the better"
+        purchasetext: "The deeper the knowledge the better",
+        type: "basic",
     },
 
     {
@@ -217,7 +235,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("magicContainer").classList.remove("hidden"); },
-        purchasetext: "FIREBALL!!!!!"
+        purchasetext: "FIREBALL!!!!!",
+        type: "basic",
     },
     {
         name: "Magic Study",
@@ -229,7 +248,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("magicstudyContainer").classList.remove("hidden"); },
-        purchasetext: "Books are boring but magic is not."
+        purchasetext: "Books are boring but magic is not.",
+        type: "basic",
     },
     {
         name: "Magic Learner",
@@ -241,7 +261,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("magiclearnerContainer").classList.remove("hidden"); },
-        purchasetext: "Time to increase learning with magic!."
+        purchasetext: "Time to increase learning with magic!.",
+        type: "basic",
     },
     {
         name: "Magic Multitasker",
@@ -253,7 +274,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("magicmultitaskerContainer").classList.remove("hidden"); },
-        purchasetext: "Learn multiple times over...somehow?"
+        purchasetext: "Learn multiple times over...somehow?",
+        type: "basic",
     },
     
     {
@@ -266,7 +288,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.querySelector('[data-tab = "Potion"]').classList.remove("hidden"); },
-        purchasetext: "Potion man giving you a call. Better head down."
+        purchasetext: "Potion man giving you a call. Better head down.",
+        type: "basic",
     },
 
     {
@@ -279,7 +302,8 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("abyssalContainer").classList.remove("hidden"); },
-        purchasetext: "The abyss resides inside the pit."
+        purchasetext: "The abyss resides inside the pit.",
+        type: "basic",
     },
 
     {
@@ -291,8 +315,9 @@ const upgrades = [
         flavortext: "A whisper echoes from around you. 'If you build it, he will come.'",
         purchased: 0,
         maxpurchases: 1,
-        onpurchase: () => { document.querySelector('[data-tab = "Energy"]').classList.remove("hidden"); },
-        purchasetext: "A well materializes in front of you, filled with green goo."
+        onpurchase: () => { document.getElementById("energyContainer").classList.remove("hidden") },
+        purchasetext: "A well materializes in front of you, filled with green goo.",
+        type: "basic",
     },
 
     {
@@ -305,7 +330,38 @@ const upgrades = [
         purchased: 0,
         maxpurchases: 1,
         onpurchase: () => { document.getElementById("insightContainer").classList.remove("hidden"); },
-        purchasetext: "Inner peace is achieved by looking at the bigger picture."
+        purchasetext: "Inner peace is achieved by looking at the bigger picture.",
+        type: "basic",
+    },
+
+    {
+        name: "Unlock Mining/Barcoins",
+        id: "unlockmining",
+        cost: [["matter", 50]],
+        unlocked: false,
+        reqs: [["barcoins", 100]],
+        flavortext: "Down by a bar, an old man told you bout some miners.",
+        purchased: 0,
+        maxpurchases: 1,
+        onpurchase: () => {
+            document.querySelector('[data-tab = "Mining"]').classList.remove("hidden");
+            document.getElementById("displayBarcoins").classList.remove("hidden");},
+        purchasetext: "Finally you go down to the big ole' mountain nearby, and discover a whole new place to explore...",
+        type: "basic",
+    },
+
+    {
+        name: "Unlock Mineshaft 1: Basic tunnel",
+        id: "unlockshaft1",
+        cost: [["matter", 50]],
+        unlocked: false,
+        reqs: [["barcoins", 1]],
+        flavortext: "Dig out the first mineshaft out of many more.",
+        purchased: 0,
+        maxpurchases: 1,
+        onpurchase: () => {},
+        purchasetext: "Finally completed.",
+        type: "mining",
     },
 ]
 
@@ -322,7 +378,7 @@ setInterval(() => {
 
             if (loop.unlocked === false) {
 
-                if (reqType === "knowledge" || reqType === "wisdom" || reqType === "matter" || reqType === "energy") {
+                if (reqType === "knowledge" || reqType === "wisdom" || reqType === "matter" || reqType === "energy" || reqType === "barcoins") {
                     if (!(player[reqType] >= reqAmount)) {
                         allReqsMet = false;
                     }
@@ -387,7 +443,14 @@ function buyUpgrade(upgradeId) {
 for (const btns of upgrades) {
     const button = document.createElement("button") // Creates buttons for upgrades
     button.id = btns.id
-    document.getElementById("upgradeContainer").appendChild(button);
+    if (btns.type === "basic") {
+        button.style.backgroundColor = `#6a5409`
+        document.getElementById("upgradeContainer").appendChild(button);
+    } else if (btns.type === "mining") {
+        button.style.backgroundColor = `#24387e`
+        document.getElementById("miningUpgradeContainer").appendChild(button);
+    }
+
 
     let allcosts = "";
     let deleteAnd = false;
@@ -443,7 +506,7 @@ document.getElementById("unlocker").addEventListener("click", () => {
 
                 }
                 console.log(eachReq[0])
-                if (eachReq[0] === "knowledge" || eachReq[0] === "wisdom" || eachReq[0] === "matter" || eachReq[0] === "energy") {
+                if (eachReq[0] === "knowledge" || eachReq[0] === "wisdom" || eachReq[0] === "matter" || eachReq[0] === "energy" || eachReq[0] === "barcoins") {
                     allreqs += `${eachReq[1]} ${eachReq[0]}`
                 } else {
                     allreqs += ` ${barInfo[eachReq[0]][1]} at ${eachReq[1]} Levels`